# backend/docs/swagger.yaml
basePath: /
definitions:
  handler.AppointmentResponse:
    properties:
      carId:
        type: string
      clientName:
        type: string
      createdAt:
        type: string
      date:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      notes:
        type: string
      service:
        type: string
      status:
        type: string
      time:
        type: string
      updatedAt:
        type: string
    type: object
...
  handler.CreateAppointmentRequest:
    properties:
      carID:
        description: ✅ camelCase
        type: string
      customerID:
        description: ✅ camelCase
        type: string
      employeeID:
        description: ✅ camelCase
        type: string
      notes:
        type: string
      reason:
        type: string
      scheduledAt:
        description: ✅ camelCase
        type: string
      scheduledTime:
        description: ✅ camelCase
        type: string
      serviceType:
        description: ✅ camelCase
        type: string
      status:
        type: string
    type: object
...
  handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.SwaggerLoginOK:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  handler.SwaggerMeOK:
    properties:
      user:
        $ref: '#/definitions/handler.SwaggerMeUser'
    type: object
...
  handler.SwaggerMessage:
    properties:
      details:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
  handler.SwaggerRegisterOK:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/handler.SwaggerRegisterUser'
    type: object
...
  handler.UpdateAppointmentRequest:
    properties:
      carId:
        description: ✅ camelCase
        type: string
      customerID:
        type: string
      employeeID:
        type: string
      notes:
        type: string
      reason:
        type: string
      scheduledAt:
        description: ✅ camelCase
        type: string
      scheduledTime:
        description: ✅ camelCase
        type: string
      serviceType:
        description: ✅ camelCase
        type: string
      status:
        type: string
    type: object
...
  ports.RegisterRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        minLength: 6
        type: string
      role:
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'API de gestión de taller: autenticación JWT, coches, citas y empleados.'
  title: GonsGarage API
  version: "1.0"
paths:
  /api/v1/appointments:
    get:
      parameters:
      - description: Filtro UUID cliente (staff)
        in: query
        name: customerId
        type: string
      - description: Filtro UUID coche
        in: query
        name: carId
        type: string
      - description: Estado (scheduled, confirmed, ...)
        in: query
        name: status
        type: string
      - description: Límite (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: created_at o scheduled_at
        in: query
        name: sortBy
        type: string
      - description: ASC o DESC
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.AppointmentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Listar citas
      tags:
      - appointments
    post:
      consumes:
      - application/json
      parameters:
      - description: carID, scheduledAt (RFC3339), serviceType, etc.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAppointmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.AppointmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Crear cita
      tags:
      - appointments
  /api/v1/appointments/{id}:
    delete:
      parameters:
      - description: UUID de la cita
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Sin cuerpo
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Eliminar cita
      tags:
      - appointments
    get:
      parameters:
      - description: UUID de la cita
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AppointmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Obtener cita
      tags:
      - appointments
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID de la cita
        in: path
        name: id
        required: true
        type: string
      - description: Campos a fusionar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateAppointmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AppointmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Actualizar cita
      tags:
      - appointments
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Devuelve un JWT (campo token). El cliente debe llamar después a
        GET /auth/me para el perfil completo.
      parameters:
      - description: Credenciales
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SwaggerLoginOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      summary: Iniciar sesión
      tags:
      - auth
  /api/v1/auth/me:
    get:
      description: Requiere cabecera Authorization con esquema Bearer y el JWT.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SwaggerMeOK'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Perfil actual
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Registro público. Email único; 409 si el usuario ya existe.
      parameters:
      - description: Datos de registro
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ports.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.SwaggerRegisterOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      summary: Registro
      tags:
      - auth
  /api/v1/cars:
    get:
      parameters:
      - description: UUID del cliente dueño (solo staff)
        in: query
        name: ownerId
        type: string
      - description: Límite (staff sin ownerId; default 50)
        in: query
        name: limit
        type: integer
      - description: Offset paginación
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.CarResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Listar coches
      tags:
      - cars
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del vehículo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCarRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Crear coche
      tags:
      - cars
  /api/v1/cars/{id}:
    delete:
      parameters:
      - description: UUID del coche
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Sin cuerpo
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Eliminar coche
      tags:
      - cars
    get:
      parameters:
      - description: UUID del coche
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Obtener coche
      tags:
      - cars
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID del coche
        in: path
        name: id
        required: true
        type: string
      - description: Campos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateCarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Actualizar coche
      tags:
      - cars
  /api/v1/employees:
    get:
      parameters:
      - description: Límite (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Filtro departamento
        in: query
        name: department
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: employees, total, limit, offset
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Listar empleados
      tags:
      - employees
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del empleado
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ports.CreateEmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: message y employee
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Crear empleado
      tags:
      - employees
  /api/v1/employees/{id}:
    delete:
      parameters:
      - description: ID del empleado
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Eliminar empleado
      tags:
      - employees
    get:
      parameters:
      - description: ID del empleado
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: employee
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Obtener empleado
      tags:
      - employees
    put:
      consumes:
      - application/json
      parameters:
      - description: ID del empleado
        in: path
        name: id
        required: true
        type: string
      - description: Campos
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ports.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message y employee
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Actualizar empleado
      tags:
      - employees
...
securityDefinitions:
  BearerAuth:
    description: 'JWT: cabecera Authorization con valor Bearer seguido del token (rutas
      bajo /api/v1 salvo /health).'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"